+++ tencentcloud/faceid/v20180301/client.go
@@ -773,7 +773,7 @@ func NewEncryptedPhoneVerificationResponse() (response *EncryptedPhoneVerificati
 }
 
 // EncryptedPhoneVerification
-// 本接口用于校验手机号、姓名和身份证号的真实性和一致性，入参支持MD5加密传输。
+// 本接口用于校验手机号、姓名和身份证号的真实性和一致性，入参支持明文、MD5和SHA256加密传输。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_DECRYPTSYSTEMERROR = "FailedOperation.DecryptSystemError"
@@ -791,7 +791,7 @@ func (c *Client) EncryptedPhoneVerification(request *EncryptedPhoneVerificationR
 }
 
 // EncryptedPhoneVerification
-// 本接口用于校验手机号、姓名和身份证号的真实性和一致性，入参支持MD5加密传输。
+// 本接口用于校验手机号、姓名和身份证号的真实性和一致性，入参支持明文、MD5和SHA256加密传输。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_DECRYPTSYSTEMERROR = "FailedOperation.DecryptSystemError"