+++ tencentcloud/trtc/v20190722/models.go
@@ -1750,6 +1750,6 @@ type WaterMarkParams struct {
 	// 水印在输出时的Y偏移。单位为像素值。
 	LocationY *uint64 `json:"LocationY,omitempty" name:"LocationY"`
 
-	// 混流-水印图片URL地址，支持png、jpg、jpeg、bmp格式，暂不支持透明通道。URL链接长度限制为512字节。WaterMarkUrl和WaterMarkId参数都填时，以WaterMarkUrl为准。图片大小限制不超过10MB。
+	// 混流-水印图片URL地址，支持png、jpg、jpeg、bmp格式，暂不支持透明通道。URL链接长度限制为512字节。WaterMarkUrl和WaterMarkId参数都填时，以WaterMarkUrl为准。图片大小限制不超过2MB。
 	WaterMarkUrl *string `json:"WaterMarkUrl,omitempty" name:"WaterMarkUrl"`
 }